refactoring and a faster DefaultCategoryDataSet for plotting in GSEA
