+++ src/main/java/org/reactome/cytoscape/genescore/GeneScoreDistributionPlotPane.java
@@ -4,12 +4,19 @@ import java.awt.BorderLayout;
 import java.awt.Color;
 import java.awt.Font;
 import java.awt.geom.Rectangle2D;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
 import javax.swing.BorderFactory;
+import javax.swing.JFrame;
 import javax.swing.JPanel;
 
+import org.gk.util.GKApplicationUtilities;
 import org.jfree.chart.ChartFactory;
 import org.jfree.chart.ChartPanel;
 import org.jfree.chart.JFreeChart;
@@ -21,6 +28,7 @@ import org.jfree.chart.renderer.category.CategoryItemRendererState;
 import org.jfree.chart.renderer.category.LayeredBarRenderer;
 import org.jfree.data.category.DefaultCategoryDataset;
 import org.jfree.data.general.DatasetChangeEvent;
+import org.reactome.r3.util.MathUtilities;
 
 /**
  * This class is used to show the ranks of a selected pathway in the whole distribution
@@ -58,9 +66,7 @@ public class GeneScoreDistributionPlotPane extends JPanel {
         geneToScore.keySet()
                    .stream()
                    .sorted((g1, g2) -> geneToScore.get(g2).compareTo(geneToScore.get(g1)))
-                   .forEach(gene -> {
-                       dataset.addValue(geneToScore.get(gene), "All Genes", gene);
-                   });
+                   .forEach(gene -> dataset.addValue(geneToScore.get(gene), "All Genes", gene));
         pathwayGenes.stream()
                     .filter(gene -> geneToScore.get(gene) != null)
                     .sorted((g1, g2) -> geneToScore.get(g2).compareTo(geneToScore.get(g1)))
@@ -72,7 +78,7 @@ public class GeneScoreDistributionPlotPane extends JPanel {
     }
 
     private JPanel createGraphPane() {
-        dataset = new DefaultCategoryDataset();
+        dataset = new HashedCategoryDataSet();
         dataset.setNotify(false);
         
         JFreeChart chart = ChartFactory.createBarChart("Gene Score Distribution",
@@ -119,9 +125,186 @@ public class GeneScoreDistributionPlotPane extends JPanel {
         plot.setDatasetRenderingOrder(DatasetRenderingOrder.REVERSE);
         
         ChartPanel chartPane = new ChartPanel(chart);
+        chartPane.setDoubleBuffered(true);
         chartPane.setBorder(BorderFactory.createEtchedBorder());
 
         return chartPane;
+    }
+    
+    /**
+     * Customized DefaultCategoryDataset class to quick performance by using Map, instead of List,
+     * for columns, which are genes. 
+     * @author wug
+     *
+     */
+    private class HashedCategoryDataSet extends DefaultCategoryDataset {
+        private Map<Comparable, Integer> columnKeyToIndex;
+        private List<Comparable> columnKeys;
+        private List<Comparable> rowKeys;
+        private Map<Comparable, Map<Comparable, Number>> rowToColToValue;
+        
+        public HashedCategoryDataSet() {
+            columnKeyToIndex = new HashMap<>();
+            rowKeys = new ArrayList<>();
+            columnKeys = new ArrayList<>();
+            rowToColToValue = new HashMap<>();
+        }
+        
+        @Override
+        public int getRowCount() {
+            return rowKeys.size();
+        }
+
+        @Override
+        public int getColumnCount() {
+            return columnKeys.size();
+        }
+
+        @Override
+        public Number getValue(int row, int column) {
+            Comparable rowKey = getRowKey(row);
+            Comparable colKey = getColumnKey(column);
+            return getValue(rowKey, colKey);
+        }
+
+        @Override
+        public Comparable getRowKey(int row) {
+            return rowKeys.get(row);
+        }
+
+        @Override
+        public int getRowIndex(Comparable key) {
+            return rowKeys.indexOf(key); // We expect to see two rows. Therefore, this should be very quick.
+        }
+
+        @Override
+        public List getRowKeys() {
+            return rowKeys;
+        }
+
+        @Override
+        public Comparable getColumnKey(int column) {
+            return columnKeys.get(column);
+        }
+
+        @Override
+        public List getColumnKeys() {
+            return columnKeys;
+        }
+
+        @Override
+        public Number getValue(Comparable rowKey, Comparable columnKey) {
+            Map<Comparable, Number> colToValue = rowToColToValue.get(rowKey);
+            if (colToValue == null)
+                return null;
+            return colToValue.get(columnKey);
+        }
+
+        @Override
+        public void addValue(Number value, Comparable rowKey, Comparable columnKey) {
+            rowToColToValue.compute(rowKey, (key, map) -> {
+                if (map == null)
+                    map = new HashMap<>();
+                map.put(columnKey, value.doubleValue());
+                return map;
+            });
+            if (!rowKeys.contains(rowKey))
+                rowKeys.add(rowKey);
+            if (columnKeyToIndex.containsKey(columnKey))
+                return;
+            columnKeyToIndex.put(columnKey, columnKeyToIndex.size() - 1);
+            columnKeys.add(columnKey);
+            fireDatasetChanged();
+        }
+
+        @Override
+        public void removeValue(Comparable rowKey, Comparable columnKey) {
+            Map<Comparable, Number> colToValue = rowToColToValue.get(rowKey);
+            if (colToValue != null)
+                colToValue.remove(columnKey);
+            fireDatasetChanged();
+        }
+
+        @Override
+        public void removeRow(int rowIndex) {
+            Comparable rowKey = getRowKey(rowIndex);
+            removeRow(rowKey);
+        }
+
+        @Override
+        public void removeRow(Comparable rowKey) {
+            rowToColToValue.remove(rowKey);
+            fireDatasetChanged();
+        }
+
+        @Override
+        public void removeColumn(int columnIndex) {
+            removeColumn(getColumnKey(columnIndex));
+        }
+
+        @Override
+        public void removeColumn(Comparable columnKey) {
+            rowToColToValue.forEach((row, map) -> map.remove(columnKey));
+            fireDatasetChanged();
+        }
+
+        @Override
+        public void clear() {
+            columnKeys.clear();
+            rowKeys.clear();
+            columnKeyToIndex.clear();
+            rowToColToValue.clear();
+        }
+
+        @Override
+        public int hashCode() {
+            return rowToColToValue.hashCode();
+        }
+
+        @Override
+        public Object clone() throws CloneNotSupportedException {
+            throw new CloneNotSupportedException("Don't clone this customized DefaultCategoryDataSet!");
+        }
+
+        @Override
+        public int getColumnIndex(Comparable key) {
+            Integer index = columnKeyToIndex.get(key);
+            if (index == null)
+                return -1;
+            return index;
+        }
+        
+    }
+    
+    /**
+     * This is a test method.
+     * @param args
+     * @throws Exception
+     */
+    public static void main(String[] args) throws Exception {
+        String fileName = "/Users/wug/Documents/eclipse_workspace/ohsu/results/beataml/Trametinib_Corr_Gene_Auc_101918.txt";
+        Map<String, Double> geneToScore = new HashMap<>();
+        Files.lines(Paths.get(fileName))
+             .skip(1)
+             .forEach(line -> {
+                 String[] tokens = line.split("\t");
+                 geneToScore.put(tokens[0], new Double(tokens[1]));
+             });
+        
+        geneToScore.keySet()
+        .stream()
+        .sorted((g1, g2) -> geneToScore.get(g2).compareTo(geneToScore.get(g1)))
+        .forEach(gene -> geneToScore.get(gene));
+        
+        GeneScoreDistributionPlotPane plotPane = new GeneScoreDistributionPlotPane();
+        Set<String> genes = MathUtilities.randomSampling(geneToScore.keySet(), 250);
+        plotPane.setGeneToScore(geneToScore, genes);
+        JFrame frame = new JFrame("Gene Score Distribution");
+        frame.getContentPane().add(plotPane, BorderLayout.CENTER);
+        frame.setSize(400, 800);
+        GKApplicationUtilities.center(frame);
+        frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+        frame.setVisible(true);
     }
 
 }
